- Added sentinel value for file path in encoder
Alright, seem it was my mistake to missed out the
sentinel value for the file path, so this is the
reason why I can easily get that does not exists
error when I creating the encoder. The file path
now throws a compile error when the users don't
create a string with sentinel termination.

Fortunately, for other types, including Decoder which
was my recent addition, contains the termination,
so other functions should throw an compile error.